refactored omics.py to use blueprints
